started working on user-auth again
